Create Proxy_Toggle
After days of trying and with the help from Kacey Coughlin, I finally managed to build a proxy toggle for mac 10.7.5 to turn GAE proxy tool on and off. I haven't try these codes on other version of mac. I wish somebody could tell me how to modified it to run this code on Mac 10.9